requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.10.6
cssselect2==0.8.0
dj-database-url==3.0.1
Django==4.2
django-celery-results==2.5.1
django-cors-headers==4.5.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==37.6.0
flower==2.0.1
fonttools==4.58.1
humanize==4.12.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
kombu==5.4.2
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pyphen==0.17.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-factoryboy==2.8.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.1
redis==6.4.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.2
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1